+++ app/scripts/api/snapshot/Transporter.js
@@ -26,6 +26,7 @@ define(function(require) {
     var SimCapiMessage = require('./SimCapiMessage');
     var check = require('check');
     var SimCapiValue = require('./SimCapiValue');
+    var SimCapiTypes = require('./SimCapiTypes');
     var ApiInterface = require('./ApiInterface');
     var LocalData = require('./LocalData');
     var domainUtil = require('./util/domain');
@@ -726,6 +727,10 @@ define(function(require) {
             if (overwriteValue !== undefined) {
                 simCapiValue.setValue(overwriteValue);
                 callChangeListeners([simCapiValue]);
+
+                 if (overwriteValue instanceof Array && (simCapiValue.type === SimCapiTypes.TYPES.ARRAY || simCapiValue.type === SimCapiTypes.TYPES.ARRAY_POINT)) {
+                     simCapiValue.value = simCapiValue.value.toString();
+                 }
             }
 
             outgoingMap[key] = simCapiValue;