+++ test/specs/api/snapshot/Transporter.spec.js
@@ -1316,6 +1316,54 @@ define(function(require) {
             });
         });
 
+        describe('Exposing a value that was waiting to be applied', function() {
+            it('should stringify the array before sending the message to the platform', function() {
+                var toBeAppliedCapiArray = new SimCapiValue({
+                    key: 'arrayProp',
+                    type: SimCapiTypes.TYPES.ARRAY,
+                    value: [1,2,3]
+                });
+                
+                var toBeApplied = {};
+                toBeApplied[toBeAppliedCapiArray.key] = toBeAppliedCapiArray.value;
+                
+                var transporter = new Transporter({
+                    authToken: 'token',
+                    toBeApplied: toBeApplied
+                });
+                sandbox.stub(transporter, 'sendMessage');
+                
+                transporter.expose(toBeAppliedCapiArray);
+                clock.tick(25);
+
+                var changedValues = transporter.sendMessage.getCall(0).args[0].values;
+                expect(changedValues[toBeAppliedCapiArray.key].value).to.equal(toBeAppliedCapiArray.value.toString());
+            });
+
+            it('should not change a value that is already a string containing an array', function() {
+                var toBeAppliedCapiArray = new SimCapiValue({
+                    key: 'arrayProp',
+                    type: SimCapiTypes.TYPES.ARRAY,
+                    value: '[foo bye,bar,baz]'
+                });
+
+                var toBeApplied = {};
+                toBeApplied[toBeAppliedCapiArray.key] = toBeAppliedCapiArray.value;
+
+                var transporter = new Transporter({
+                    authToken: 'token',
+                    toBeApplied: toBeApplied
+                });
+                sandbox.stub(transporter, 'sendMessage');
+
+                transporter.expose(toBeAppliedCapiArray);
+                clock.tick(25);
+
+                var changedValues = transporter.sendMessage.getCall(0).args[0].values;
+                expect(changedValues[toBeAppliedCapiArray.key].value).to.equal(toBeAppliedCapiArray.value);
+            });
+        });
+
         describe('RESIZE_PARENT_CONTAINER request and response', function() {
             var options, onSuccess;
             beforeEach(function() {